refactoring under progress
