+++ lib/main.dart
@@ -6,7 +6,7 @@ import 'dart:io';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  AtSignLogger.root_level = 'finer';
+  // AtSignLogger.root_level = 'finer';
   if (Platform.isLinux || Platform.isMacOS) {
     await DesktopWindow.setWindowSize(Size(1200, 700));
     await DesktopWindow.setMinWindowSize(Size(1200, 700));
